hello_tutorial: Corrected type for _id.

+++ repos/hello_tutorial/src/hello/server/main.cc
@@ -28,7 +28,7 @@ namespace Hello {
 
 struct Hello::Session_component : Genode::Rpc_object<Session>
 {
-	unsigned int _id;
+	unsigned short _id;
 
 	Session_component(unsigned short id) : Genode::Rpc_object<Session>(), _id(id) {}
